+++ src/LinkSharer.php
@@ -11,6 +11,7 @@ namespace MichaelHall\LinkSharer;
 use DataTypes\Interfaces\UrlInterface;
 use MichaelHall\LinkSharer\Sharers\FacebookSharer;
 use MichaelHall\LinkSharer\Sharers\GooglePlusSharer;
+use MichaelHall\LinkSharer\Sharers\LinkedInSharer;
 use MichaelHall\LinkSharer\Sharers\TwitterSharer;
 
 /**
@@ -58,6 +59,18 @@ class LinkSharer
     public function getGooglePlusSharer(): GooglePlusSharer
     {
         return new GooglePlusSharer($this->url);
+    }
+
+    /**
+     * Returns the LinkedIn sharer.
+     *
+     * @since 1.0.0
+     *
+     * @return LinkedInSharer The LinkedIn sharer.
+     */
+    public function getLinkedInSharer(): LinkedInSharer
+    {
+        return new LinkedInSharer($this->url, $this->text);
     }
 
     /**