+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Added
+- LinkedIn sharer class.
 
 ## 1.0.0 - 2017-06-05
 - First stable revision.